Fix onboarding end date not saving — cast $3 to int in SQL
The parameterized query `$2::date + $3` failed silently because
PostgreSQL couldn't resolve `date + unknown`. Fixed by casting
to `$3::int`. This was why onboarding created a start event but
no end event, leaving the period showing as eternally active.

Also cleaned up Kaylee's data: replaced orphaned March 1 start
with proper start/end pair (March 1–6).

+++ src/app/api/kids/health/route.ts
@@ -605,7 +605,7 @@ export async function POST(request: NextRequest) {
           )
           // Insert end event at start + duration
           await query(
-            `INSERT INTO kid_cycle_log (kid_name, event_type, event_date) VALUES ($1, 'end', $2::date + $3)`,
+            `INSERT INTO kid_cycle_log (kid_name, event_type, event_date) VALUES ($1, 'end', $2::date + $3::int)`,
             [childLower, lastPeriodStart, dur]
           )
         }
